implement models and corresponding db migrations
Models contain the Blogpost as persisted in the database and the BlogspotRequest in the form it is recieved.

The request will contain the avatar iamge as a url string and the post image as a base64 string.

The server would  store both images in the file system then persist the coprresponding filenames in the database.